use signed ints to compare rs optimes in stepdown

+++ db/repl/replset_commands.cpp
@@ -236,10 +236,10 @@ namespace mongo {
             // only step down if there is another node synced to within 10
             // seconds of this node
             if (!force) {
-                unsigned lastOp = theReplSet->lastOpTimeWritten.getSecs();
-                unsigned closest = theReplSet->lastOtherOpTime().getSecs();
+                long long int lastOp = (long long int)theReplSet->lastOpTimeWritten.getSecs();
+                long long int closest = (long long int)theReplSet->lastOtherOpTime().getSecs();
 
-                int diff = lastOp - closest;
+                long long int diff = lastOp - closest;
                 result.append("closest", closest);
                 result.append("difference", diff);
